\name{ANN}
\alias{ANN-package}
\alias{Artificial Neural Network}
\docType{package}
\title{
Artificial Neural Network
}
\description{
A simple artificial neural network (ANN) implementation for classification. Contains three core methods; ANN.train(), ANN.validate(), ANN.classify()
}
\details{
\tabular{ll}{
Package: \tab ANN\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-12-30\cr
License: \tab MIT\cr
}
ANN.train()
ANN.validate()
ANN.classify()
}
\author{
Author: Benjamin Cordier
Maintainer: Benjamin Cordier (benacordier(at)gmail.com)
}
\references{

}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ 
package
neuralnet
artificial
neural
network
}
\examples{
See example.r
}
