\name{ANN}
\alias{plot.ANN}
\alias{print.ANN}
\title{Printing/plot an ANN object}
\description{
Printing/plot method for an ANN objects}

\usage{
  \method{print}{ANN}(x,...)
  \method{plot}{ANN}(x,type="p",...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{Object of class "ANN", an already optimized ANN.}
  \item{type}{type of plot "p" for points, "l" for lines, etc.}
  \item{\dots}{not used}
}
%\details{
%  to be continued...
%}
%\value{
%%  ~Describe the value returned
% If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
%}
%\references{
%put references to the literature/web site here 
%}
\author{
Francis Roy-Desrosiesrs
}
%\note{
%further notes
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

%\seealso{

%% ~~objects to See Also as \code{\link{help}}, ~~~
%}
