\name{demoResultANN}
\docType{data}
\alias{demoResultANN}
\alias{fLog}
\alias{input}
\alias{noisy_sin}
\alias{output}
\alias{t}
\title{Result of the demo}
\description{
List of objects created by the demo. Increases accessibilty to results by eliminating computation step.
}
\usage{demoResultANN}
\examples{
data("dataANN")
demoResultANN
}
\keyword{datasets}


